Update Rudecli_GUI-tkinter.py
CTCP now ignores any requests no in the allowed list. 
Updated all file and folder pointers to the absolute path of the script. This way you can run the script from the full path or from the current working directory and have it not matter. 
Updated cowsay to have cow be closer to the speech bubble.